refactor: split POST stats and recommendations (#4899)
